Bugfixes and blocked commit

+++ menu.py
@@ -1,14 +1,14 @@
-from sqlalchemy import String, Integer, Column, Table, ForeignKey
+from sqlalchemy import String, Integer, Column, Table, ForeignKey, Float
 from sqlalchemy.orm import relationship
 
-from base_template import Base
+from base_template import Base, Session
 
 class Menu(Base):
     __tablename__ = 'menu'
 
     id = Column(Integer, primary_key = True)
     name = Column(String)
-    price = Column(Integer)
+    price = Column(Float)
     recipe = relationship('Stock')
 
     def __init__(self, name, price):
